+++ src/common/input.ts
@@ -1709,13 +1709,20 @@ export class Input extends EventEmitter {
             case 'notify':
             case 'not':
                 if (args.length === 0)
-                    throw new Error('Invalid syntax use \x1b[4m#not\x1b[0;-11;-12mify title message');
+                    throw new Error('Invalid syntax use \x1b[4m#not\x1b[0;-11;-12mify title \x1b[3mmessage icon\x1b[0;-11;-12m');
                 else {
                     args[0] = this.stripQuotes(args[0]);
+                    n = { silent: true };
+                    if (args[args.length - 1].match(/^\{.*\}$/g)) {
+                        item = args.pop();
+                        n.icon = parseTemplate(this.parseOutgoing(item.substr(1, item.length - 2), false));
+                    }
+                    if (args.length === 0)
+                        throw new Error('Invalid syntax use \x1b[4m#not\x1b[0;-11;-12mify title \x1b[3mmessage icon\x1b[0;-11;-12m');
                     if (args.length === 1)
-                        this.client.notify(this.parseOutgoing(args[0], false), null);
+                        this.client.notify(this.parseOutgoing(this.stripQuotes(args[0]), false), null, n);
                     else
-                        this.client.notify(this.parseOutgoing(args[0], false), this.parseOutgoing(args.slice(1).join(' '), false));
+                        this.client.notify(this.parseOutgoing(this.stripQuotes(args[0]), false), this.parseOutgoing(args.slice(1).join(' '), false), n);
                 }
                 return null;
             case 'idle':